src/phonebook/pyproject.toml
[project]
name = "phonebook"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-dotenv>=1.0",
    "PyJWT[crypto]>=2.8",
    "jwcrypto>=1.5",
]
